cgofuse: support `nofuse` build tag to disable pkg
This should assure that the fuse dependency isn't even fetched during
build, let alone required to build. Runtime stubs are added so that
callers may depend on the interface (at compile time), without requiring
it to be usable (at runtime).
